Search based on location

+++ controllers/reviews.js
@@ -1,24 +1,33 @@
 const Listing = require("../models/listing");
 const Review = require("../models/review");
 
-module.exports.createReview = async(req,res)=>{
- let listing = await Listing.findById(req.params.id);
- let newReview = new Review (req.body.review);
- console.log(newReview);
- newReview.author= req.user._id;
- listing.reviews.push(newReview);
+module.exports.createReview = async (req, res) => {
+    let listing = await Listing.findById(req.params.id).populate("reviews");
+    let newReview = new Review(req.body.review);
+    newReview.author = req.user._id;
 
- await newReview.save();
- await listing.save();
- req.flash("success","New Review Created");
- res.redirect(`/listings/${listing._id}`);
+    listing.reviews.push(newReview);
+
+    await newReview.save();
+    await listing.save();
+
+    req.flash("success", "New Review Created");
+    res.redirect(`/listings/${listing._id}`);
 };
 
-module.exports.destroyReview = async(req,res)=>{
-  let {id,reviewId} = req.params;
+module.exports.destroyReview = async (req, res) => {
+    let { id, reviewId } = req.params;
+    let listing = await Listing.findById(id);
+    let review = await Review.findById(reviewId);
 
-  await Listing.findByIdAndUpdate(id,{$pull:{reviews:reviewId}});
-  await Review.findByIdAndDelete(reviewId);
-  req.flash("success","Review Deleted");
-  res.redirect(`/listings/${id}`);
+    // Check if the user is the review author or the listing owner
+    if (review.author.equals(req.user._id) || listing.owner.equals(req.user._id)) {
+        await Listing.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
+        await Review.findByIdAndDelete(reviewId);
+        req.flash("success", "Review Deleted");
+    } else {
+        req.flash("error", "You do not have permission to delete this review.");
+    }
+
+    res.redirect(`/listings/${id}`);
 };
